pyproject.toml
[project]
name = "walmart-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "faster-whisper>=1.1.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "supabase>=2.16.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.2",
    "uvicorn>=0.35.0",
]
